requirements.txt
sentence-transformers
pdfplumber
pinecone
python-dotenv
tqdm

openai
tiktoken
pydantic

streamlit
bcrypt
pymongo
pymongo